requirements.txt
osmnx>=1.5.1
matplotlib>=3.5.0
geopandas>=0.13.0
pyyaml>=6.0
pyrosm>=0.6.0
requests>=2.31.0
shapely>=2.0.0
